gaussian_splatting: extract streaming diagnostics surface into its own TU (decomposition slice 3)
Stacked on slices 1 (layout-hint) and 2 (chunk-invariants). Moves the eight
diagnostics-reporting/aggregation member methods of GaussianStreamingSystem
out of gaussian_streaming.cpp into a new partial-class translation unit
core/streaming_diagnostics_surface.cpp (the proven ISSUE-006 partial-class
pattern; declarations stay in gaussian_streaming.h):
  _build_streaming_diagnostics_snapshot, get_streaming_analytics,
  get_task_debug_state, get_chunk_culling_stats, get_vram_debug_stats,
  _log_streaming_telemetry, _log_streaming_frame_stats, end_frame.
begin_frame stays (frame lifecycle, not diagnostics).

Prerequisite, same change: two anon-namespace queue-pressure helper wrappers
(_summarize_queue_pressure_checked, _validate_queue_pressure_latched_state)
were called by both the moved methods and a surviving method, so they were
relocated to their natural owner — StreamingQueuePressureController
(streaming_queue_pressure_controller.{h,cpp}) — as static methods, and all
call sites re-qualified. This both unblocks the move and puts the helpers
where they belong.

Result: gaussian_streaming.cpp 5463 -> 4558 LOC (6051 -> 4558 across slices
1+2+3, ~25% of the original god-file now lives in focused units). Pure code
move, behavior-preserving (logic byte-identical; only the ClassDB bind_method
pointers for the bound getters remain in gaussian_streaming.cpp, resolving to
the relocated definitions via partial-class linkage).

Risk: R2 (streaming). Evidence: full editor build of the combined slice1+2+3
tree compiles + links clean (scons ... dev_build=yes tests=yes, exit 0; only
pre-existing C4458 warnings in gpu_buffer_manager.h); guards green
(run_module_tests.py --guard-only, 108 OK).
